Renamed additional references from ng2-select to ng-next-select. Included migration section in readme
