+++ angular-cli.json
@@ -1,6 +1,6 @@
 {
   "project": {
-    "name": "ng2-select"
+    "name": "ng-next-select"
   },
   "apps": [
     {
