+++ demo/src/index.html
@@ -14,7 +14,7 @@
   <!--link to bootstrap.css-->
   <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css">
   <link rel="stylesheet" href="assets/css/style.css">
-  <link rel="stylesheet" href="assets/css/ng2-select.css">
+  <link rel="stylesheet" href="assets/css/ng-next-select.css">
   <link rel="stylesheet" href="assets/css/prettify-angulario.css">
 
   <style media="screen">
