+++ src/tsconfig.json
@@ -28,7 +28,7 @@
   ],
   "files": [
     "../scripts/typings.d.ts",
-    "./ng2-select.ts",
+    "./ng-next-select.ts",
     "./index.ts"
   ],
   "angularCompilerOptions": {
